+++ src/raws/raws.hpp
@@ -14,7 +14,7 @@ struct string_table_t {
 	std::vector<std::string> strings;
 
 	inline std::string random_entry(rltk::random_number_generator rng=rltk::random_number_generator()) {
-		const int position = rng.roll_dice(1, strings.size()) - 1;
+		const int position = rng.roll_dice(1, static_cast<int>(strings.size())) - 1;
 		return strings[position];
 	}
 };
